requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0

# Production WSGI server (required for Heroku)
gunicorn==21.2.0
eventlet==0.33.3

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
